+++ backend/src/main/java/anyone/to/soma/user/application/UserService.java
@@ -5,8 +5,8 @@ import anyone.to.soma.exception.repository.NoSuchRecordException;
 import anyone.to.soma.user.domain.Profile;
 import anyone.to.soma.user.domain.User;
 import anyone.to.soma.user.domain.UserRepository;
-import anyone.to.soma.user.dto.LoginResponse;
-import anyone.to.soma.user.dto.ProfileRequest;
+import anyone.to.soma.user.domain.dto.LoginResponse;
+import anyone.to.soma.user.domain.dto.ProfileRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
