+++ backend/src/test/java/anyone/to/soma/user/UserServiceMockTest.java
@@ -9,8 +9,8 @@ import anyone.to.soma.user.domain.type.Age;
 import anyone.to.soma.user.domain.type.Gender;
 import anyone.to.soma.user.domain.type.Job;
 import anyone.to.soma.user.domain.type.LoginType;
-import anyone.to.soma.user.dto.LoginResponse;
-import anyone.to.soma.user.dto.ProfileRequest;
+import anyone.to.soma.user.domain.dto.LoginResponse;
+import anyone.to.soma.user.domain.dto.ProfileRequest;
 import anyone.to.soma.utils.Fixtures;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
