+++ backend/src/test/java/anyone/to/soma/user/UserServiceTest.java
@@ -10,8 +10,8 @@ import anyone.to.soma.user.domain.UserRepository;
 import anyone.to.soma.user.domain.type.Age;
 import anyone.to.soma.user.domain.type.Gender;
 import anyone.to.soma.user.domain.type.Job;
-import anyone.to.soma.user.dto.LoginResponse;
-import anyone.to.soma.user.dto.ProfileRequest;
+import anyone.to.soma.user.domain.dto.LoginResponse;
+import anyone.to.soma.user.domain.dto.ProfileRequest;
 import anyone.to.soma.utils.LetterDecorationRepository;
 import anyone.to.soma.utils.ReplyRepository;
 import com.auth0.jwt.interfaces.DecodedJWT;
